+++ src/source/lexer.rs
@@ -66,6 +66,7 @@ impl Lexer {
                     | TokenKind::Slash
                     | TokenKind::Colon
                     | TokenKind::At
+                    | TokenKind::Caret
                     | TokenKind::Dot
             )
         ) || self.next_starts_infix()
@@ -77,7 +78,7 @@ impl Lexer {
         }
         matches!(
             self.input.get(index),
-            Some('+' | '-' | '*' | '/' | '<' | '>' | '=' | '!' | '&' | '|' | '@' | '.')
+            Some('+' | '-' | '*' | '/' | '<' | '>' | '=' | '!' | '&' | '|' | '@' | '^' | '.',)
         )
     }
     fn push(tokens: &mut Vec<Token>, kind: TokenKind, span: SourceSpan) {
@@ -213,6 +214,10 @@ impl Lexer {
                 '@' => {
                     self.next();
                     Self::push(&mut result, TokenKind::At, span);
+                }
+                '^' => {
+                    self.next();
+                    Self::push(&mut result, TokenKind::Caret, span);
                 }
                 '.' => {
                     self.next();